update failed ui-tests
